Remove outdated test file for ModularCheckout and consolidate related test cases into a single describe block for improved organization and clarity. Enhance Apple Pay event handling tests and ensure proper error emission when no payment method is selected.
